[#49] Rename View to Container for Map component
