+++ src/components/Map/index.js
@@ -4,4 +4,4 @@ export { default as TileLayer } from './TileLayer';
 export { default as GeoJson } from './GeoJson';
 export { default as LegendControl } from './LegendControl';
 export { default as MarkerClusterGroup } from './MarkerClusterGroup';
-export { default as View } from './View';
+export { default as Container } from './Container';
